feat: fix delete story

+++ lib/app/features/chat/e2ee/providers/send_chat_message/send_e2ee_chat_message_service.c.dart
@@ -146,7 +146,7 @@ class SendE2eeChatMessageService {
         return a.compareTo(b);
       });
 
-      if (mediaAttachmentsUsersBased.isEmpty && content.isEmpty) {
+      if (mediaAttachmentsUsersBased.isEmpty && content.isEmpty && quotedEvent == null) {
         await ref.read(eventMessageDaoProvider).deleteByEventReference(eventReference);
         return sentMessage;
       }
